Created a component for handling login and authentication
Seperated this into two separate files. lock.ts handles the interaction with auth0. We expose an authenticated stream which will fire events when authentication is triggered on or off. Additionally, functions for triggering login, logout, and getting the current token, if any, are exposed.

The login component is simply a react component that uses the lock logic to show a login/logout button and toggle between them based on the auth state.